Switch to 23S17 from 23017
This increases the scan rate by ~5x (from 1.7MHz clock to 10MHz) and also
works around an erratum (?) in the 23017 which forces the A7 and B7 gpios
to be outputs.  This could lead to a short circuit if (for example)
a key on the same SENSE line is held down in both the A7 and some other
row -- A7 would be driven high (because it can't be three-stated) while
the other row is driven low, and a direct connection exists through
the two key switches.

The 23S17 part requires a few extra I/O pins, but we have some to spare.
Disconnected the INT pin from the I/O expander, because the way we would
do an interrupt is by driving all low on the 23S17 pins and then watching
for a low on the RP2040 GPIOs.  We shouldn't need to trigger an interrupt
on the DRV lines from the I/O expander.  But I pulled the INT line out to
a test point just in case.